Initial commit for scc-hypervisor-collector
This defines a reasonable starting infrastructure for the project with
a functional CLI command, and an initial version of the ConfigManager
that will read on YAML config files.

Packaging is implemented by the provided setup.py and setup.cfg files.

Partial pytest based testing is implemented in the tests/ directory.

A tox.ini is provided that can be used to drive both code quality
and compliance checks, with support for testing with multiple Python
versions is pyenv is configured appropriately.

The bin/ directory contains helper scripts that are useful for any
developers working on the project, which can be used to create a
Python virtual env with the recommended tools installed, and some
wrapper scripts that can be used to run those commands from the
venv, or if not found there, from the system environment.

TODO:
  * Implement 90% coverage via pytest tests.
  * Re-enable pytest failure reporting in tox runs once coverage
    goal met.
  * Document configuration schema appropriately.
  * Implement configuration schema validation.
  * Implement hypervisor data collection support.
  * Implement collection scheduling.
  * Implement SCC data upload once new SCC API is available.