support check mode to report all possible config errors
The check mode is used to process and report all the issues
found with the configuration instead of failing on the first
config issue. We maintain the list of config errors in
self._config_errors and append to the list as we encounter
configuration error while processing the configuration.
valid is now recursive as we check validity of the child
config instances.

Fixes #5