update handling of the parsed backends key
when you load a yaml file with an empty backends:
section, you get None assigned to the dict key
backends which was causing the copy() in the
_merge_config_data method to fail with
"AttributeError: 'NoneType' object has no attribute 'copy'".
This change fixes the issue by assigning an empty list
if the backends section is None.
Also added a test for this scenario.

Fixes #19

other fixes in this change include:
  fail early when empty backends section is specified.
  tests updated accordingly