Add helper scripts for the tox dev venv
Our tox.ini defines a dev test that sets up a venv with all of the
required dependencies installed as well as installing our tool in
developer mode.

Add some helper scripts to the bin directory to leverage this to
better enable local adhoc testing of the tool.

Update the README.md to reflect these changes and add some details
about the tox dev venv. Also update the tox testing overview to
align with with recent changes to the tox.ini. Move the CI status
image to the top of the README.md, and add a leading exclamation
mark to indicate that we want to show the referenced image.