check-manifest needs to ignore spec file.
Add the OBS spec file to list of files that should be ignored by the
check-manifest tool as part of the tox 'check' test.

Also move check-manifest ignore option specification to be managed by
setup.cfg rather than as a option on the check-manifest command line
in the tox commands list.

Relates: #11