Tests for checking logs for sensitive information
Introduced a new fixture to create the hypervisor collector
with different log levels and exercise the test against
these different log levels to validate we are not leaking
any sensitive information. We set the log level
of the root in the fixture so all the modules
inherit the log level being tested against.

Fixes #8