systemd settings tweaks and add OBS container build support.
Add support for building a container with the tool installed in the
openSUSE Build Service.

Also in the systemd service run the scc-hypervisor-collector using
the timeout command, with a 1 hour timeout, to ensure that a hanging
network connection doesn't permanently stall the scheduling of the
service.

Switch the systemd timer script to trigger running the associated
service on a daily basis.

Update README.md to reflect recent packaging related changes.

Relates: #11
Fixes: #34