Skip CLI credentials check if NO_NETWORK_ACCESS env var set
The build service doesn't permit external network access so we need
to skip the SCC credentials check CLI test when building the package.

The spec file sets the NO_NETWORK_ACCESS env var to true before it
runs the pytests, and the pytests check if that env var is set, and
skip the test if needed.