Implement file save/upload mechanisms
Add support for a collect and save to file mode, with a corresponding
upload from previously saved file mode.

To collect the hypervisor details and save them to a file, the user
can leverage the new '--output' option to specify an output file in
which the collected details for all backends will be saved, in YAML
format.

To upload (previously collected) data from a file the new '--input'
option can be used to specify the YAML file containing the saved
collection details to the SCC.

Note that the uploading system will not be required to specify the
backends associated with the details being uploaded as part of the
scc-hypervisor-collection config settings; only the SCC credentials
config will be required on the uploading system.

However the system doing the collection still needs SCC creds to
be specified, but these will not be used, and potentiall can be
dummy values.

Summary of Code changes:

* Modified the HypervisorCollector.details to be a separate class
  (HypervisorDetails) hosting the collected details in the required
  payload format for uploading to the SCC.
* Added a new CollectionResults class to host the details that have
  been collected by the various hypervisors managed by the scheduler,
  that can be initialised from a provided scheduler, or can save and
  load details to or from a YAML file.
* Modified the CLI '--upload' option handling to leverage the new
  CollectionScheduler.results property (CollectionResults) as the'
  source for data to be uploaded.
* Added CLI support for the '--output' option which saves the new
  CollectionScheduler.results property's contents to a file, and
  skips the upload step.
* Added CLI support for the '--input' option which loads the YAML
  file content as the collection results which are used to upload
  the data to the SCC, skipping the scheduler driven collection of
  data from configured backends altogether.
* Split the permissions checking static method out of ConfigManager
  into the api.util subpackage, so that it could be leveraged by
  the CollectionResults load & save methods.
* Misc changes to code to address pylint complaints caused by the
  addition of code to existing complex functions/methods that lead
  to them exceeding complexity thresholds.
* Added unit testing for the new features implemented in this patch
  as well as some additional tests for the api.uploader subpackage
  to increase our coverage back over 90%.