Make SCC server url configurable
To aid in testing against development environment instances of the SCC,
allow the target end point to be configured is the SCC credentials.

Added unit tests to validate that the default value is returned when
no url value is specified in the config, and that the provided value
is returned when a url value is specified in the config.

This change can be backed out in the long run if desired.