Updated comments to clarify what a test is doing.
For the CollectionResults.load testing we are creating new temp
files and ensuring that the file permissions are set appropriately
to exercise the desired test scenario, e.g. correct ownership and
permissions allowing a load() to complete successfully, or invalid
permissions or ownership to trigger the desired exception which the
test is seeking to catch.