Prototype for automated version tagging as a workflow
This workflow is triggered to run when a PR is closed because it was
merged, and will automatically run bumpversion to increment the patch
level in the version if the version value wasn't changed as part of
the PR. Finally it will create a tag for the new version, whether it
was updated as part of the PR, or by running bumpversion.

These changes required the disabling of bumpversion's automatic tag
creating functionality.

NOTE: This mechanism only works for PRs based upon branches within
the main repo. It will not work for PRs created for forks of the
main repo.